+++ common/lib/xmodule/xmodule/partitions/partitions_service.py
@@ -9,12 +9,11 @@ import logging
 
 import six
 from django.conf import settings
-from django.utils.translation import ugettext_lazy as _
 
 from openedx.core.lib.cache_utils import request_cached
-from openedx.features.content_type_gating.partitions import create_content_gating_partition
+from openedx.core.lib.dynamic_partitions_generators import DynamicPartitionGeneratorsPluginManager
 from xmodule.modulestore.django import modulestore
-from xmodule.partitions.partitions import ENROLLMENT_TRACK_PARTITION_ID, UserPartition, UserPartitionError
+from xmodule.partitions.partitions import get_partition_from_id
 
 log = logging.getLogger(__name__)
 
@@ -74,48 +73,14 @@ def _get_dynamic_partitions(course):
     Return the dynamic user partitions for this course.
     If none exists, returns an empty array.
     """
-    return [
-        partition
-        for partition in [
-            _create_enrollment_track_partition(course),
-            create_content_gating_partition(course),
-        ]
-        if partition
-    ]
+    dynamic_partition_generators = DynamicPartitionGeneratorsPluginManager.get_available_plugins().values()
+    generated_partitions = []
+    for generator in dynamic_partition_generators:
+        generated_partition = generator(course)
+        if generated_partition:
+            generated_partitions.append(generated_partition)
 
-
-def _create_enrollment_track_partition(course):
-    """
-    Create and return the dynamic enrollment track user partition.
-    If it cannot be created, None is returned.
-    """
-    if not FEATURES.get('ENABLE_ENROLLMENT_TRACK_USER_PARTITION'):
-        return None
-
-    try:
-        enrollment_track_scheme = UserPartition.get_scheme("enrollment_track")
-    except UserPartitionError:
-        log.warning("No 'enrollment_track' scheme registered, EnrollmentTrackUserPartition will not be created.")
-        return None
-
-    used_ids = set(p.id for p in course.user_partitions)
-    if ENROLLMENT_TRACK_PARTITION_ID in used_ids:
-        log.warning(
-            "Can't add 'enrollment_track' partition, as ID {id} is assigned to {partition} in course {course}.".format(
-                id=ENROLLMENT_TRACK_PARTITION_ID,
-                partition=get_partition_from_id(course.user_partitions, ENROLLMENT_TRACK_PARTITION_ID).name,
-                course=six.text_type(course.id)
-            )
-        )
-        return None
-
-    partition = enrollment_track_scheme.create_user_partition(
-        id=ENROLLMENT_TRACK_PARTITION_ID,
-        name=_(u"Enrollment Track Groups"),
-        description=_(u"Partition for segmenting users by enrollment track"),
-        parameters={"course_id": six.text_type(course.id)}
-    )
-    return partition
+    return generated_partitions
 
 
 class PartitionService(object):
@@ -203,17 +168,3 @@ class PartitionService(object):
         return user_partition.scheme.get_group_for_user(
             self._course_id, user, user_partition, assign=assign,
         )
-
-
-def get_partition_from_id(partitions, user_partition_id):
-    """
-    Look for a user partition with a matching id in the provided list of partitions.
-
-    Returns:
-        A UserPartition, or None if not found.
-    """
-    for partition in partitions:
-        if partition.id == user_partition_id:
-            return partition
-
-    return None
