+++ setup.py
@@ -108,5 +108,9 @@ setup(
         'openedx.learning_context': [
             'lib = openedx.core.djangoapps.content_libraries.library_context:LibraryContextImpl',
         ],
+        'openedx.dynamic_partition_generator': [
+            'enrollment_track = common.lib.xmodule.xmodule.partitions.enrollment_track_partition_generator:create_enrollment_track_partition',
+            'content_type_gating = openedx.features.content_type_gating.partitions:create_content_gating_partition'
+        ],
     }
 )
